+++ changelog.md
@@ -17,6 +17,8 @@
 
 ### fix
 - saving ahk lib in wrong location.
+- taskbar not been restored when changing weg enabled state.
+- taskbar been restored always as not autohide, now it will restored as initial state.
 
 ## [1.10.2]
 ### fix